python/kafka-consumer/requirements.txt
requests==2.31.0
kafka-python==2.0.2
